requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
click==8.1.8
fastapi==0.115.12
flake8==7.2.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.13.0
pydantic==2.11.3
pydantic-core==2.33.1
pyflakes==3.3.2
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.0
uvicorn==0.34.2
websockets==15.0.1

# Added dependencies
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
psutil==5.9.8
